+++ src/test.rs
@@ -1,6 +1,9 @@
 #[cfg(test)]
 pub mod helpers {
+    use crate::config;
+    use crate::config::Config;
     use crate::items::{DateInfo, Item};
+    use std::collections::HashMap;
 
     /// Checks if environment supports colored output (GitHub Actions does not)
     pub fn supports_coloured_output() -> bool {
@@ -22,9 +25,21 @@ pub mod helpers {
             is_deleted: 0,
         }
     }
+
+    pub fn config_fixture() -> Config {
+        Config {
+            token: String::from("alreadycreated"),
+            projects: HashMap::new(),
+            path: config::generate_path().unwrap(),
+            next_id: None,
+            last_version_check: None,
+            timezone: Some(String::from("US/Pacific")),
+        }
+    }
 }
 #[cfg(test)]
 pub mod responses {
+    use crate::test::helpers;
     use crate::{time, VERSION};
 
     pub fn sync() -> String {
@@ -75,7 +90,7 @@ pub mod responses {
                 }}
             ]
         }}",
-            time::today_string()
+            time::today_string(&helpers::config_fixture())
         )
     }